Long to BigInteger

+++ src/main/java/com/code/monks/csm/dto/analytics/projection/CustomerAmountView.java
@@ -1,7 +1,9 @@
 package com.code.monks.csm.dto.analytics.projection;
 
+import java.math.BigInteger;
+
 public interface CustomerAmountView {
 
     Integer getCustomerId();
-    Long getAmount();
+    BigInteger getAmount();
 }
