+++ src/main/java/com/code/monks/csm/dto/analytics/projection/MonthlyAnalyticsView.java
@@ -1,9 +1,11 @@
 package com.code.monks.csm.dto.analytics.projection;
 
+import java.math.BigInteger;
+
 public interface MonthlyAnalyticsView {
 
     Integer getYear();
     Integer getMonth();
-    Long getAmount();
+    BigInteger getAmount();
     Long getCount();
 }
